pyproject.toml
[tool.poetry]
name = "dbworkload"
version = "0.10.2"
description = "Workload framework"
authors = ["Fabio Ghirardello"]
license = "GPLv3+"
readme = "README.md"
homepage = "https://dbworkload.github.io/dbworkload/"
repository = "https://github.com/dbworkload/dbworkload"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent"
]

[tool.poetry.scripts]
dbworkload = 'dbworkload.cli.main:app'

[tool.poetry.dependencies]
python = "^3.11"
pandas = "*"
tabulate = "*"
numpy = "*"
prometheus-client = "*"
pyyaml = "*"
typer = {extras = ["all"], version = "*"}
cassandra-driver = {version = "*", optional = true}
mysql-connector-python = {version = "*", optional = true}
oracledb = {version = "*", optional = true}
psycopg = {version = "*", optional = true}
psycopg-binary = {version = "*", optional = true}
pymongo = {version = "*", optional = true}
pyodbc = {version = "*", optional = true}
mariadb = {version = "*", optional = true}
google-cloud-spanner = {version = "*", optional = true}
pinecone = {version = "^8.0.0", optional = true, extras = ["asyncio", "grpc"]}
sentence-transformers = { version = "*", optional = true }
langchain-core = { version = "^1.0.5", optional = true }
langchain-openai = { version = "^1.0.2", optional = true }
langchain-ollama = { version = "^1.0.0", optional = true }
fastembed = {version = "^0.7.3", optional = true }
pgvector = {version = "^0.4.1", optional = true }
langgraph = { version = "^1.0.3", optional = true }
openai = { version = "^2.8.0", optional = true }
pytdigest = "*"
plotext = "*"
plotly = "*"
jinja2 = "*"
sqlparse = "*"
psutil = "^7.0.0"

[tool.poetry.extras]
all =  ["psycopg", "psycopg-binary", "mysql-connector-python", "mariadb", "oracledb", "pyodbc", "pymongo", "cassandra-driver", "google-cloud-spanner", "pinecone", "convert"]
postgres = ["psycopg", "psycopg-binary"]
mysql = ["mysql-connector-python"]
mariadb = ["mariadb"]
oracle = ["oracledb"]
odbc = ["pyodbc"]
mongo = ["pymongo"]
cassandra = ["cassandra-driver"]
spanner = ["google-cloud-spanner"]
pinecone = ["pinecone", "sentence-transformers"]
convert = [
  "langchain-core",
  "langchain-openai",
  "langchain-ollama",
  "langgraph",
  "openai",
  "fastembed",
  "pgvector"
]

[tool.poetry.group.dev.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.6.14"
mkdocs-click = "^0.9.0"
watchdog = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
